# Improve MultipartReader performance