--- a/src/Http/WebUtilities/src/BufferedReadStream.cs
+++ b/src/Http/WebUtilities/src/BufferedReadStream.cs
@@ -352,11 +352,7 @@ public class BufferedReadStream : Stream
 
             while (!foundCRLF && EnsureBuffered())
             {
-                if (builder.Length > lengthLimit)
-                {
-                    throw new InvalidDataException($"Line length limit {lengthLimit} exceeded.");
-                }
-                ProcessLineChar(builder, ref foundCR, ref foundCRLF);
+                ProcessLineChar(builder, lengthLimit, ref foundCR, ref foundCRLF);
             }
 
             return DecodeLine(builder, foundCRLF);
@@ -380,30 +376,33 @@ public class BufferedReadStream : Stream
 
             while (!foundCRLF && await EnsureBufferedAsync(cancellationToken))
             {
-                if (builder.Length > lengthLimit)
-                {
-                    throw new InvalidDataException($"Line length limit {lengthLimit} exceeded.");
-                }
-
-                ProcessLineChar(builder, ref foundCR, ref foundCRLF);
+                ProcessLineChar(builder, lengthLimit, ref foundCR, ref foundCRLF);
             }
 
             return DecodeLine(builder, foundCRLF);
         }
     }
 
-    private void ProcessLineChar(MemoryStream builder, ref bool foundCR, ref bool foundCRLF)
+    private void ProcessLineChar(MemoryStream builder, int lengthLimit,  ref bool foundCR, ref bool foundCRLF)
     {
-        var b = _buffer[_bufferOffset];
-        builder.WriteByte(b);
-        _bufferOffset++;
-        _bufferCount--;
-        if (b == LF && foundCR)
+        while (_bufferCount > 0)
         {
-            foundCRLF = true;
-            return;
+            var b = _buffer[_bufferOffset];
+            builder.WriteByte(b);
+            _bufferOffset++;
+            _bufferCount--;
+            if (b == LF && foundCR)
+            {
+                foundCRLF = true;
+                return;
+            }
+            foundCR = b == CR;
+
+            if (builder.Length > lengthLimit)
+            {
+                throw new InvalidDataException($"Line length limit {lengthLimit} exceeded.");
+            }
         }
-        foundCR = b == CR;
     }
 
     private static string DecodeLine(MemoryStream builder, bool foundCRLF)