@davidfowl 2023-10-17
>  instead of writing one byte at a time.

😢